# feat(optimize): working Powell, check results are near optimum