## Description

The scirs2-optimize minimize tests did not check the results properly so I edited them to get a better idea of what was working. This PR supplies these fixed tests, which show the following tests not passing:

```bash
cargo nextest run -E "test(minimize)" -p scirs2-optimize
```

```
     Summary [   0.004s] 11 tests run: 6 passed, 5 failed, 12 skipped
        FAIL [   0.002s] scirs2-optimize unconstrained::tests::test_minimize_bfgs_rosenbrock
        FAIL [   0.003s] scirs2-optimize unconstrained::tests::test_minimize_cg_rosenbrock
        FAIL [   0.003s] scirs2-optimize unconstrained::tests::test_minimize_nelder_mead_rosenbrock
        FAIL [   0.002s] scirs2-optimize unconstrained::tests::test_minimize_powell_quadratic
        FAIL [   0.002s] scirs2-optimize unconstrained::tests::test_minimize_powell_rosenbrock
```

<details><summary>Click to show details</summary>

```
    Finished `test` profile [optimized + debuginfo] target(s) in 0.10s
────────────
 Nextest run ID 5d7992bf-8246-485f-9ec2-e5956ed2e3d9 with nextest profile: default
    Starting 11 tests across 1 binary (12 tests skipped)
        PASS [   0.002s] scirs2-optimize unconstrained::tests::test_minimize_nelder_mead_quadratic
        PASS [   0.002s] scirs2-optimize constrained::tests::test_minimize_constrained_placeholder
        FAIL [   0.002s] scirs2-optimize unconstrained::tests::test_minimize_cg_rosenbrock
──── STDOUT:             scirs2-optimize unconstrained::tests::test_minimize_cg_rosenbrock

running 1 test
test unconstrained::tests::test_minimize_cg_rosenbrock ... FAILED

failures:

failures:
    unconstrained::tests::test_minimize_cg_rosenbrock

test result: FAILED. 0 passed; 1 failed; 0 ignored; 0 measured; 22 filtered out; finished in 0.00s

──── STDERR:             scirs2-optimize unconstrained::tests::test_minimize_cg_rosenbrock

thread 'unconstrained::tests::test_minimize_cg_rosenbrock' panicked at scirs2-optimize/src/unconstrained.rs:959:9:
assert_relative_eq!(result.fun, 1.0, epsilon = 1e-6)

    left  = 0.9531265749391484
    right = 1.0


note: run with `RUST_BACKTRACE=1` environment variable to display a backtrace

  Cancelling due to test failure: 8 tests still running
        FAIL [   0.002s] scirs2-optimize unconstrained::tests::test_minimize_nelder_mead_rosenbrock
──── STDOUT:             scirs2-optimize unconstrained::tests::test_minimize_nelder_mead_rosenbrock

running 1 test
test unconstrained::tests::test_minimize_nelder_mead_rosenbrock ... FAILED

failures:

failures:
    unconstrained::tests::test_minimize_nelder_mead_rosenbrock

test result: FAILED. 0 passed; 1 failed; 0 ignored; 0 measured; 22 filtered out; finished in 0.00s

──── STDERR:             scirs2-optimize unconstrained::tests::test_minimize_nelder_mead_rosenbrock

thread 'unconstrained::tests::test_minimize_nelder_mead_rosenbrock' panicked at scirs2-optimize/src/unconstrained.rs:861:9:
assert_relative_eq!(result.fun, 1.0, epsilon = 1e-6)

    left  = 1.63627702139686e-9
    right = 1.0


note: run with `RUST_BACKTRACE=1` environment variable to display a backtrace

        FAIL [   0.002s] scirs2-optimize unconstrained::tests::test_minimize_powell_rosenbrock
──── STDOUT:             scirs2-optimize unconstrained::tests::test_minimize_powell_rosenbrock

running 1 test
test unconstrained::tests::test_minimize_powell_rosenbrock ... FAILED

failures:

failures:
    unconstrained::tests::test_minimize_powell_rosenbrock

test result: FAILED. 0 passed; 1 failed; 0 ignored; 0 measured; 22 filtered out; finished in 0.00s

──── STDERR:             scirs2-optimize unconstrained::tests::test_minimize_powell_rosenbrock

thread 'unconstrained::tests::test_minimize_powell_rosenbrock' panicked at scirs2-optimize/src/unconstrained.rs:911:9:
assert_relative_eq!(result.fun, 1.0, epsilon = 1e-6)

    left  = 2.5370588688189625e-15
    right = 1.0


note: run with `RUST_BACKTRACE=1` environment variable to display a backtrace

        PASS [   0.002s] scirs2-optimize constrained::tests::test_minimize_trust_constr
        PASS [   0.003s] scirs2-optimize unconstrained::tests::test_minimize_bfgs_quadratic
        PASS [   0.003s] scirs2-optimize constrained::tests::test_minimize_slsqp
        FAIL [   0.003s] scirs2-optimize unconstrained::tests::test_minimize_bfgs_rosenbrock
──── STDOUT:             scirs2-optimize unconstrained::tests::test_minimize_bfgs_rosenbrock

running 1 test
test unconstrained::tests::test_minimize_bfgs_rosenbrock ... FAILED

failures:

failures:
    unconstrained::tests::test_minimize_bfgs_rosenbrock

test result: FAILED. 0 passed; 1 failed; 0 ignored; 0 measured; 22 filtered out; finished in 0.00s

──── STDERR:             scirs2-optimize unconstrained::tests::test_minimize_bfgs_rosenbrock

thread 'unconstrained::tests::test_minimize_bfgs_rosenbrock' panicked at scirs2-optimize/src/unconstrained.rs:828:9:
assert_relative_eq!(result.fun, 1.0, epsilon = 1e-6)

    left  = 0.09630967484392935
    right = 1.0


note: run with `RUST_BACKTRACE=1` environment variable to display a backtrace

        PASS [   0.003s] scirs2-optimize unconstrained::tests::test_minimize_cg_quadratic
        FAIL [   0.003s] scirs2-optimize unconstrained::tests::test_minimize_powell_quadratic
──── STDOUT:             scirs2-optimize unconstrained::tests::test_minimize_powell_quadratic

running 1 test
test unconstrained::tests::test_minimize_powell_quadratic ... FAILED

failures:

failures:
    unconstrained::tests::test_minimize_powell_quadratic

test result: FAILED. 0 passed; 1 failed; 0 ignored; 0 measured; 22 filtered out; finished in 0.00s

──── STDERR:             scirs2-optimize unconstrained::tests::test_minimize_powell_quadratic

thread 'unconstrained::tests::test_minimize_powell_quadratic' panicked at scirs2-optimize/src/unconstrained.rs:879:9:
assert_relative_eq!(result.x[0], 0.0, epsilon = 1e-2)

    left  = 1.0000330534806758
    right = 0.0


note: run with `RUST_BACKTRACE=1` environment variable to display a backtrace

────────────
     Summary [   0.004s] 11 tests run: 6 passed, 5 failed, 12 skipped
        FAIL [   0.003s] scirs2-optimize unconstrained::tests::test_minimize_bfgs_rosenbrock
        FAIL [   0.002s] scirs2-optimize unconstrained::tests::test_minimize_cg_rosenbrock
        FAIL [   0.002s] scirs2-optimize unconstrained::tests::test_minimize_nelder_mead_rosenbrock
        FAIL [   0.003s] scirs2-optimize unconstrained::tests::test_minimize_powell_quadratic
        FAIL [   0.002s] scirs2-optimize unconstrained::tests::test_minimize_powell_rosenbrock
error: test run failed
```

</details>

## Type of change

- [x] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to not work as expected)
- [ ] This change requires a documentation update
- [ ] AI/ML related functionality

## How has this been tested?
<!-- Please describe the tests that you ran to verify your changes -->
- [x] Unit tests
- [ ] Integration tests
- [ ] Performance tests (if relevant)

## Checklist
- [x] My code follows the style guidelines of this project
- [x] I have performed a self-review of my own code
- [x] I have commented my code, particularly in hard-to-understand areas
- [ ] I have made corresponding changes to the documentation
- [ ] My changes generate no new warnings
- [x] I have added tests that prove my fix is effective or that my feature works
- :no_entry: New and existing unit tests pass locally with my changes
- [ ] Any dependent changes have been merged and published in downstream modules

## Additional context

- Update: Powell quadratic is fixed now